Readded Texture Mapping
I figure it might be neat to have texture mapping, so I added it back shortly after removing it. The code associated with it is pretty janky, at times, so I was tempted to cut it out to avoid showing off what I consider to be bad code, but it is neat enough to be worth implementing.